MAIN: some refactor with no functional changes
piped/generate_commit_message.go:

- Removed unused imports: `os/user`, `path/filepath`.
- Refactored `expandTilde` function to `utils.ExpandTilde` and `isValidPath` function to `utils.IsValidPath`.
- Refactored `isGitRepository` function to `utils.IsGitRepository`.
- Renamed `JsonWithComments` to `ChangedFilesMapWithComments`.
- Moved the logic of joining file paths and comments into a separate function called `joinCommentsToString`.
- Updated the usage of `iterateJSONAndMarkChanges` to use `ChangedFilesMapWithComments`.
- Updated the output file with the formatted result string using `utils.AnswersOutput`.
There were some comments added to the code.

utils/utils.go:

- Added `os/user` and `path/filepath` import statements.
- Added `IsGitRepository` function to check if a given folder path is a git repository.
- Added `ExpandTilde` function to expand a path with a tilde (~) to the user's home directory.
- Added `isFolderPath` function to check if a given path is a folder.
- Added `IsValidPath` function to check if a given path is a valid folder path.
- Updated error handling in `IfFileWithAPiKeyExists` function.
